🚚Rename: 파일,폴더 구조 다시 세팅 (#48)
- 폴더 구조를 src로 뺄 수 있는 것들을 빼고 layout분리 하였습니다.
- 타입 alias 를 설정하였습니다.
- 어쩌다보니 tailwind prettier 에러를 해결했습니다.
