+++ src/app/(unauthorized)/layout.tsx
@@ -1,9 +1,7 @@
-import AuthLayout from '@ui/layout/AuthLayout';
-
 export default function UnAuthorizedLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  return <AuthLayout>{children}</AuthLayout>;
+  return <>{children}</>;
 }
