--- src/app/Provider/QueryProvider.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import { QueryClientProvider } from '@tanstack/react-query';
-import { queryClient } from '../api/queryClient';
-
-export default function QueryProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  );
-}
