--- src/app/api/queryClient.ts
@@ -1,3 +1,0 @@
-import { QueryClient } from '@tanstack/react-query';
-
-export const queryClient = new QueryClient();
