+++ src/app/layout.tsx
@@ -1,10 +1,8 @@
 import type { Metadata } from 'next';
 import './globals.css';
-import QueryProvider from './Provider/QueryProvider';
 import { SpeedInsights } from '@vercel/speed-insights/next';
-import MainLayout from './layout/MainLayout';
-import Popup from './ui/common/Popup';
-
+import QueryProvider from '@provider/QueryProvider';
+import Popup from '@ui/common/Popup';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -20,9 +18,8 @@ export default function RootLayout({
     <QueryProvider>
       <html lang="en">
         <body>
-          <MainLayout>{children}</MainLayout>
-          <Popup />
           {children}
+          <Popup />
           <SpeedInsights />
         </body>
       </html>
