--- src/app/lib/api/axios.ts
@@ -1,24 +1,0 @@
-import axios from 'axios';
-
-export const instance = axios.create({
-  baseURL: '',
-  timeout: 10000,
-});
-
-axios.interceptors.request.use(
-  function (config) {
-    return config;
-  },
-  function (error) {
-    return Promise.reject(error);
-  },
-);
-
-axios.interceptors.response.use(
-  function (response) {
-    return response;
-  },
-  function (error) {
-    return Promise.reject(error);
-  },
-);
