--- src/app/lib/authOptions.ts
@@ -1,8 +1,0 @@
-import { NextAuthOptions } from 'next-auth';
-
-export const authOptions: NextAuthOptions = {
-  secret: 'next auth secret key',
-  providers: [],
-  pages: {},
-  callbacks: {},
-};
