--- src/app/playground/page.tsx
@@ -1,5 +1,0 @@
-'use client';
-
-export default function Playground() {
-  return <div></div>;
-}
