--- src/app/store/auth.store.ts
@@ -1,5 +1,0 @@
-import { create } from 'zustand';
-
-const useStore = create(() => ({}));
-
-export default useStore;
