--- src/app/ui/common/Header.tsx
@@ -1,3 +1,0 @@
-export default function Header() {
-  return <div>Header</div>;
-}
