--- src/types.ts
@@ -1,4 +1,0 @@
-// 타입 정의 하기
-
-// HyperLink type
-export type THyperLinkType = 'icon' | 'normal';
