+++ src/types/input.ts
@@ -1,5 +1,4 @@
 import { InputHTMLAttributes } from 'react';
-import { UseFormRegister } from 'react-hook-form';
 
 export type TInputErrorType =
   | 'required'
@@ -13,6 +12,5 @@ export interface IInputProps
   extends Omit<InputHTMLAttributes<HTMLInputElement>, 'type'>,
     React.RefAttributes<HTMLInputElement> {
   type: 'text' | 'email';
-  errorMessage?: string;
-  register?: UseFormRegister<Record<string, string>>;
+  errorType?: TInputErrorType;
 }
