+++ src/ui/common/CheckBox.tsx
@@ -15,7 +15,7 @@ export default function CheckBox(props: TCheckBoxProps) {
         id={uid}
         type="checkbox"
         className={twMerge(
-          'h-4 w-4 flex-shrink-0 cursor-pointer appearance-none rounded-md border border-slate-200 bg-white p-[.125rem]',
+          'h-4 w-4 flex-shrink-0 cursor-pointer appearance-none rounded-md border border-slate-200 p-[.125rem]',
           "checked:border-primary checked:bg-primary checked:bg-[url('/asset/image/check-icon.svg')] checked:bg-center checked:bg-no-repeat",
         )}
         checked={checked}
