+++ src/ui/common/Header.tsx
@@ -1,9 +1,3 @@
 export default function Header() {
-  return (
-    <div className="flex h-12 w-full items-center gap-4 bg-white md:hidden">
-      <button className="h-6 w-6 text-slate-400">
-        {/* 햄버거 아이콘 들어감, sidebar 여는 핸들러 등록필요함 */}=
-      </button>
-    </div>
-  );
+  return <div>Header</div>;
 }
