+++ src/ui/common/Input.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from 'react';
 import { twMerge } from 'tailwind-merge';
-import { INPUT_CLASSNAME } from '@constant/input';
+import { INPUT_CLASSNAME, INPUT_MESSAGE } from '@constant/input';
 import { IInputProps } from '@type/input';
 
 /**
@@ -11,20 +11,23 @@ import { IInputProps } from '@type/input';
  * @param {string} label - input 필드에 대한 레이블 텍스트를 제공합니다. 선택 사항입니다.
  * @returns {JSX.Element} - type에 맞는 입력 필드를 리턴합니다.
  */
+
 const Input = forwardRef<HTMLInputElement, IInputProps>((props, ref) => {
-  const { errorMessage, ...rest } = props;
+  const { errorType, ...rest } = props;
 
   const inputClassName = twMerge(
     INPUT_CLASSNAME.base,
-    errorMessage && INPUT_CLASSNAME.error.border,
+    errorType && INPUT_CLASSNAME.error.border,
   );
 
   return (
     <div className="relative pb-7">
       <input className={inputClassName} ref={ref} {...rest} />
 
-      {errorMessage && (
-        <p className={INPUT_CLASSNAME.error.message}>{errorMessage}</p>
+      {errorType && (
+        <p className={INPUT_CLASSNAME.error.message}>
+          {INPUT_MESSAGE.error[errorType]}
+        </p>
       )}
     </div>
   );
